Restore SimpleTitleOrLogo for CollectionFolderList
